+++ Anviz.SDK/Commands/SetFaceTemplateCommand.cs
@@ -9,7 +9,7 @@ namespace Anviz.SDK.Commands
         {
             var payload = new byte[15366];
             Bytes.Write(5, employeeID).CopyTo(payload, 0);
-            payload[5] = (byte)(1);
+            payload[5] = 1;
             template.CopyTo(payload, 6);
             BuildPayload(SET_FACETEMPLATE, payload);
         }
